moved networkx interops to base class
