+++ src/napari_graph/__init__.py
@@ -1,6 +1,6 @@
 from napari_graph.base_graph import BaseGraph
 from napari_graph.directed_graph import DirectedGraph
-from napari_graph.interops import to_napari_graph, to_networkx
+from napari_graph.interops import to_napari_graph
 from napari_graph.undirected_graph import UndirectedGraph
 
 try:
@@ -13,5 +13,4 @@ __all__ = [
     "DirectedGraph",
     "UndirectedGraph",
     "to_napari_graph",
-    "to_networkx",
 ]
