+++ src/napari_graph/base_graph.py
@@ -1,6 +1,9 @@
+from __future__ import annotations
+
 from abc import abstractmethod
 from typing import Callable, List, Optional, Tuple, Union
 
+import networkx as nx
 import numpy as np
 import pandas as pd
 from numba import njit, typed
@@ -772,3 +775,67 @@ class BaseGraph:
     def is_spatial(self) -> bool:
         """True if self is a spatial graph (has coordinates attribute)."""
         return self._coords is not None
+
+    @staticmethod
+    def from_networkx(graph: nx.Graph) -> BaseGraph:
+        """Loads a Directed or Undirected napari-graph from a NetworkX graph.
+
+        Parameters
+        ----------
+        graph : nx.Graph
+            The NetworkX graph to be converted.
+        """
+        from napari_graph.directed_graph import DirectedGraph
+        from napari_graph.undirected_graph import UndirectedGraph
+
+        coords_dict = nx.get_node_attributes(graph, "pos")
+        if len(coords_dict) > 0:
+            coords_df = pd.DataFrame.from_dict(coords_dict, orient="index")
+        else:
+            coords_df = None
+
+        edges = graph.edges
+        if len(edges) > 0:
+            edges = np.atleast_2d(edges)
+
+        if graph.is_directed():
+            out_graph = DirectedGraph(edges, coords_df)
+        else:
+            out_graph = UndirectedGraph(edges, coords_df)
+
+        return out_graph
+
+    def to_networkx(self) -> nx.Graph:
+        """Convert it self into NetworkX graph.
+
+        Parameters
+        ----------
+        graph : BaseGraph
+            napari-graph Graph
+
+        Returns
+        -------
+        nx.Graph
+            An equivalent NetworkX graph.
+        """
+        from napari_graph.directed_graph import DirectedGraph
+
+        if isinstance(self, DirectedGraph):
+            out_graph = nx.DiGraph()
+        else:
+            out_graph = nx.Graph()
+
+        if self.is_spatial():
+            for node_id, pos in zip(self.get_nodes(), self.get_coordinates()):
+                out_graph.add_node(node_id, pos=pos)
+        else:
+            out_graph.add_nodes_from(self.get_nodes())
+
+        edges = self.get_edges()
+        if isinstance(edges, list) and len(edges) > 0:
+            edges = np.concatenate(edges, axis=0)
+
+        edges_as_tuples = list(map(tuple, edges))
+        out_graph.add_edges_from(edges_as_tuples)
+
+        return out_graph
